+++ src/main/java/ure/render/URERenderer.java
@@ -6,10 +6,15 @@ import ure.URECamera;
 import ure.URECommander;
 
 public interface URERenderer {
+
+    public interface KeyListener {
+        void keyPressed(char key);
+    }
+
     void initialize();
     boolean windowShouldClose();
     void pollEvents();
-    void setCommander(URECommander commander); // TODO: This should be an event listener, not a direct reference to commander
+    void setKeyListener(KeyListener listener);
     int cellWidth(); // TODO: Cell width should be determined at a higher level?
     int cellHeight(); // TODO: Cell height should be determined at a higher level?
     void drawCamera(URECamera camera);
